# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Service Provider & Consumer API",
    "version": "1.0.0",
    "description": "API for managing service providers and consumers (SQLite version)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
         "/api/consumer-create": {
       "post": {
         "summary": "Create a new consumer",
         "description": "Creates a new consumer with name, city, and 6-digit PIN. Validates that the consumer doesn't already exist with the same name and city combination.",
         "tags": ["Consumers"],
         "requestBody": {
           "required": true,
           "content": {
             "application/json": {
               "schema": {
                 "$ref": "#/components/schemas/Consumer"
               },
               "examples": {
                 "validConsumer": {
                   "summary": "Valid consumer data",
                   "value": {
                     "name": "John Doe",
                     "city": "Karachi",
                     "pin": "123456"
                   }
                 }
               }
             }
           }
         },
         "responses": {
           "201": {
             "description": "Consumer created successfully",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "success": { "type": "boolean" },
                     "message": { "type": "string" },
                     "data": { "$ref": "#/components/schemas/ConsumerResponse" }
                   }
                 }
               }
             }
           },
           "400": {
             "description": "Validation error - missing or invalid fields",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "error": { "type": "string" },
                     "message": { "type": "string" }
                   }
                 }
               }
             }
           },
           "409": {
             "description": "Conflict - consumer already exists with same name and city",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "error": { "type": "string" },
                     "message": { "type": "string" }
                   }
                 }
               }
             }
           },
           "500": {
             "description": "Internal server error"
           }
         }
       }
     },
           "/api/consumer-signin": {
        "post": {
          "summary": "Sign in consumer",
          "description": "Authenticate consumer using 6-digit PIN",
         "tags": ["Consumers"],
         "requestBody": {
           "required": true,
           "content": {
             "application/json": {
               "schema": {
                 "$ref": "#/components/schemas/ConsumerSignIn"
               },
                               "examples": {
                  "signinConsumer": {
                    "summary": "Sign in credentials",
                    "value": {
                      "pin": "123456"
                    }
                  }
                }
             }
           }
         },
         "responses": {
           "200": {
             "description": "Consumer signed in successfully",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "success": { "type": "boolean" },
                     "message": { "type": "string" },
                     "data": { "$ref": "#/components/schemas/ConsumerResponse" }
                   }
                 }
               }
             }
           },
           "400": {
             "description": "Validation error - missing or invalid fields",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "error": { "type": "string" },
                     "message": { "type": "string" }
                   }
                 }
               }
             }
           },
           "401": {
             "description": "Unauthorized - invalid credentials",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "error": { "type": "string" },
                     "message": { "type": "string" }
                   }
                 }
               }
             }
           },
           "500": {
             "description": "Internal server error"
           }
         }
       }
     },
...
  },
  "components": {
    "schemas": {
...
             "Consumer": {
         "type": "object",
         "properties": {
           "name": { 
             "type": "string",
             "description": "Consumer's full name",
             "example": "John Doe"
           },
           "city": { 
             "type": "string",
             "description": "Consumer's city",
             "example": "Karachi"
           },
           "pin": { 
             "type": "string", 
             "description": "6-digit PIN code",
             "pattern": "^[0-9]{6}$",
             "minLength": 6,
             "maxLength": 6,
             "example": "123456"
           }
         },
         "required": ["name", "city", "pin"]
       },
       "ConsumerSignIn": {
         "type": "object",
         "properties": {
           "pin": { 
             "type": "string", 
             "description": "6-digit PIN code",
             "pattern": "^[0-9]{6}$",
             "minLength": 6,
             "maxLength": 6,
             "example": "123456"
           }
         },
         "required": ["pin"]
       },
      "ConsumerResponse": {
        "type": "object",
        "properties": {
          "id": { 
            "type": "integer",
            "description": "Unique consumer ID",
            "example": 1
          },
          "name": { 
            "type": "string",
            "description": "Consumer's full name",
            "example": "John Doe"
          },
          "city": { 
            "type": "string",
            "description": "Consumer's city",
            "example": "Karachi"
          },
                  "pin": {
          "type": "string",
          "description": "6-digit PIN code",
          "example": "123456"
        },
          "createdAt": { 
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": { 
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
    }
  }
}